Clean up and seperated welcome and goodbye message
